Fix Chart initialized in unmounted component (#3724)
* Fix Chart initialized in unmounted component

The `initChart` method will in some case asynchronously create a chart object (after loading `chart.js/auto`). If during the asynchronous process the Chart component was already unmounted, the Chart will still be created, but for a `null` canvas. As a consequence, a zombie Chart is created that is never cleaned up.

* Update Chart.js

Co-authored-by: Melloware <mellowaredev@gmail.com>